refactor(mealplan): relocate shopping-list writer as consumer-owned contract
Moves the shopping-list write contract out of Yumney.Shared (where it had
to use a primitive ownerId) into Yumney.MealPlan.Application.Interfaces,
where it can carry an OwnerIdentifier value object — matching the rule
established in #527.

- new IShoppingListWriter in MealPlan.Application/Interfaces (OwnerIdentifier)
- new ShoppingItemRequest DTO in MealPlan.Application/DTOs
- HttpShoppingListWriter relocated to MealPlan.Infrastructure/ExternalServices
- removed Shared.Common.IShoppingListWriter / ShoppingItemRequest
- removed orphaned in-process Shopping.Infrastructure.Services.ShoppingListWriter
  (never consumed; Shopping API uses its own AddManualItemCommand handler)
  and its dedicated tests
