+++ src/Yumney.MealPlan.Application/DTOs/ShoppingItemRequest.cs
@@ -1,0 +1,3 @@
+namespace SmartSolutionsLab.Yumney.MealPlan.Application.DTOs;
+
+public sealed record ShoppingItemRequest(string ItemName, decimal Quantity, string? Unit, string Source);
